python/pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.9.0", "pybind11", "setuptools", "wheel"]
build-backend = "scikit_build_core.build"

[project]
name = "terragen"
version = "0.1.0"
description = "Modular terrain generator with C++ noise core and Python CLI"
authors = [{ name = "Your Name" }]
requires-python = ">=3.9"
dependencies = ["numpy", "pillow", "matplotlib"]

[project.scripts]
terragen = "terragen.cli:main"

[tool.scikit-build]
wheel.abi-tag = "none"
cmake.minimum-version = "3.15"
build-dir = "build"
